Fixed some bugs
Bugs Fixed:
1.Grid disappeared in iPhone5s use ARM64
2.Force type convert problem in iPhone5s use ARM64
Codes Optimized

+++ src/CocoaChartsSample/CocoaChartsSample/CCSStringUtils.h
@@ -136,7 +136,7 @@
 - (NSString *) currency:(NSString *)code;
 
 // deci=4 123456 ->123,456.0000
-- (NSString *) decimal:(int)deci;
+- (NSString *) decimal:(NSUInteger)deci;
 
 // 123456 -> +123,456
 - (NSString *) decimalWithSign;
